feat(ai): add chatbot assistant

+++ src/main/java/org/hlopes/library/dto/SeasonProgressResponse.java
@@ -13,5 +13,5 @@ public record SeasonProgressResponse(
         LocalDate airDate,
         int episodeCount,
         boolean watched,
-        Integer rating,
+        Short rating,
         Instant watchedAt) {}
