+++ src/main/java/org/hlopes/library/dto/SeasonWatchRequest.java
@@ -1,3 +1,3 @@
 package org.hlopes.library.dto;
 
-public record SeasonWatchRequest(Integer rating) {}
+public record SeasonWatchRequest(Short rating) {}
