+++ src/main/java/org/hlopes/library/dto/SeasonWatchResponse.java
@@ -3,4 +3,4 @@ package org.hlopes.library.dto;
 import java.time.Instant;
 import java.util.UUID;
 
-public record SeasonWatchResponse(UUID id, UUID seasonId, Integer rating, Instant watchedAt) {}
+public record SeasonWatchResponse(UUID id, UUID seasonId, Short rating, Instant watchedAt) {}
